feat(ui): implement the 12 needs-review items -- tokens, IA, structure, resilience
Design-decision + shared-primitive items (stage 1):
- New semantic tokens: --color-on-accent (white-on-solid-accent, both themes),
  --color-on-warning (near-black on yellow warning pairs), --modal-backdrop
  (dialog scrim). Swept the 4 modals + sidebar drawer onto --modal-backdrop,
  text-black warning pairs onto text-on-warning.
- Removed dead --color-border-strong alias (line 41 already won the cascade).
- PageContainer now mirrors PageHeader's safe-area horizontal padding so the
  sticky header lines up on notched iOS PWA; added optional maxWidth prop (7xl
  default, 5xl/4xl) so narrow pages can adopt the scaffold.
- AnalyticsTimeFilter mode tabs go full-width/flex on phone (no overflow <400px).
- Settings Toggle gets a 44px ::before hit area; goal-form buttons min-h-11 on mobile.

Page-breadth items (stage 2, via workflow):
- PageContainer adoption across hand-rolled page roots (kept deliberate centered
  layouts like UploadSync hand-rolled).
- isError branches added where query hooks expose it but only isLoading was handled.
- KPI sizing/padding converged onto --text-kpi scale where it had drifted.
- Split oversized files: Sidebar -> BrandHeader + navConfig, SubscriptionTracker
  -> components/{AddRecurringForm,RecurringCard} + constants + types, YearInReview
  -> MonthlyBreakdownChart + YearInsightsPanel.

Fixed regressions from the earlier text-white->text-foreground sweep: 5 solid-accent
buttons (ConfirmDialog, SubscriptionTracker, Goals, HomePage, Hero) were rendering
near-black text on solid accent fills in light theme; now use text-on-accent.
Comparison KPI grid 3+1 orphan at md fixed (2x2 -> 1x4).

Gate: type-check + lint clean, 266 tests pass. Both themes verified live.
