+++ frontend/src/components/shared/ProfileModal.tsx
@@ -129,7 +129,7 @@ function ProfileModalContent({ onClose }: Readonly<{ onClose: () => void }>) {
       initial={{ opacity: 0 }}
       animate={{ opacity: 1 }}
       exit={{ opacity: 0 }}
-      className="fixed inset-0 z-50 bg-black/60 backdrop-blur-sm flex items-center justify-center p-4"
+      className="fixed inset-0 z-50 bg-[var(--modal-backdrop)] backdrop-blur-sm flex items-center justify-center p-4"
       onClick={handleClose}
     >
       <motion.div
