+++ frontend/src/pages/goals/GoalsPage.tsx
@@ -34,7 +34,7 @@ export default function GoalsPage() {
         action={
           <button
             onClick={() => state.setShowCreateForm(!state.showCreateForm)}
-            className="flex items-center gap-2 px-4 py-2 rounded-xl text-sm font-medium text-foreground transition-colors hover:opacity-90"
+            className="flex items-center gap-2 px-4 py-2 rounded-xl text-sm font-medium text-on-accent transition-colors hover:opacity-90"
             style={{ background: `linear-gradient(135deg, ${rawColors.app.blue}, ${rawColors.app.indigo})` }}
           >
             <Plus className="w-4 h-4" /> Create Goal
