+++ frontend/src/pages/goals/components/EditGoalForm.tsx
@@ -84,14 +84,14 @@ export default function EditGoalForm({
         <div className="flex gap-2">
           <button
             onClick={handleSave}
-            className="flex items-center gap-1.5 px-3 py-2 rounded-lg text-xs font-medium text-foreground transition-colors hover:opacity-90"
+            className="flex items-center justify-center gap-1.5 px-3 py-2 min-h-11 sm:min-h-0 rounded-lg text-xs font-medium text-on-accent transition-colors hover:opacity-90"
             style={{ background: `linear-gradient(135deg, ${rawColors.app.blue}, ${rawColors.app.indigo})` }}
           >
             <Save className="w-3.5 h-3.5" /> Save Changes
           </button>
           <button
             onClick={onCancel}
-            className="flex items-center gap-1.5 px-3 py-2 rounded-lg text-xs text-muted-foreground bg-[var(--overlay-2)] border border-border hover:bg-[var(--overlay-5)] transition-colors"
+            className="flex items-center justify-center gap-1.5 px-3 py-2 min-h-11 sm:min-h-0 rounded-lg text-xs text-muted-foreground bg-[var(--overlay-2)] border border-border hover:bg-[var(--overlay-5)] transition-colors"
           >
             <X className="w-3.5 h-3.5" /> Cancel
           </button>
