+++ frontend/src/pages/home/components/Hero.tsx
@@ -87,7 +87,7 @@ export function Hero({ isAuthenticated, onGetStarted, onTryDemo }: Readonly<Hero
           >
             <button
               onClick={onGetStarted}
-              className="group flex items-center gap-2 px-8 py-4 rounded-2xl font-semibold text-foreground transition-[color,background-color,border-color,transform,box-shadow] duration-300 hover:scale-105 hover:shadow-xl hover:shadow-[0_15px_40px_rgba(74,158,255,0.4)]"
+              className="group flex items-center gap-2 px-8 py-4 rounded-2xl font-semibold text-on-accent transition-[color,background-color,border-color,transform,box-shadow] duration-300 hover:scale-105 hover:shadow-xl hover:shadow-[0_15px_40px_rgba(74,158,255,0.4)]"
               style={{
                 background: `linear-gradient(135deg, ${rawColors.app.blue}, ${rawColors.app.indigo})`,
                 boxShadow: `0 10px 30px ${rawColors.app.blue}50`,
