+++ frontend/src/pages/income-expense-flow/IncomeExpenseFlowPage.tsx
@@ -1,5 +1,5 @@
 import AnalyticsTimeFilter from '@/components/shared/AnalyticsTimeFilter'
-import { PageHeader } from '@/components/ui'
+import { PageContainer, PageHeader } from '@/components/ui'
 
 import { FlowSummaryCards } from './components/FlowSummaryCards'
 import { SankeyChart } from './components/SankeyChart'
@@ -9,35 +9,33 @@ const IncomeExpenseFlowPage = () => {
   const m = useIncomeExpenseFlow()
 
   return (
-    <div className="min-h-dvh p-4 md:p-6 lg:p-8">
-      <div className="max-w-7xl mx-auto space-y-6">
-        <PageHeader
-          title="Cash Flow"
-          subtitle="Visualize how your income flows into savings and expenses"
-          action={<AnalyticsTimeFilter {...m.timeFilterProps} />}
-        />
+    <PageContainer>
+      <PageHeader
+        title="Cash Flow"
+        subtitle="Visualize how your income flows into savings and expenses"
+        action={<AnalyticsTimeFilter {...m.timeFilterProps} />}
+      />
 
-        <FlowSummaryCards
-          totalIncome={m.totalIncome}
-          totalExpense={m.totalExpense}
-          netSavings={m.netSavings}
-          savingsRate={m.savingsRate}
-        />
+      <FlowSummaryCards
+        totalIncome={m.totalIncome}
+        totalExpense={m.totalExpense}
+        netSavings={m.netSavings}
+        savingsRate={m.savingsRate}
+      />
 
-        <SankeyChart
-          isLoading={m.isLoading}
-          isMobile={m.isMobile}
-          sankeyData={m.sankeyData}
-          sankeyNodeComponent={m.sankeyNodeComponent}
-          topIncome={m.topIncome}
-          topExpense={m.topExpense}
-          totalIncome={m.totalIncome}
-          totalExpense={m.totalExpense}
-          netSavings={m.netSavings}
-          currentFY={m.currentFY}
-        />
-      </div>
-    </div>
+      <SankeyChart
+        isLoading={m.isLoading}
+        isMobile={m.isMobile}
+        sankeyData={m.sankeyData}
+        sankeyNodeComponent={m.sankeyNodeComponent}
+        topIncome={m.topIncome}
+        topExpense={m.topExpense}
+        totalIncome={m.totalIncome}
+        totalExpense={m.totalExpense}
+        netSavings={m.netSavings}
+        currentFY={m.currentFY}
+      />
+    </PageContainer>
   )
 }
