+++ frontend/src/pages/subscription-tracker/types.ts
@@ -1,0 +1,6 @@
+export interface Suggestion {
+  name: string
+  type: 'Income' | 'Expense'
+  frequency: string
+  category: string
+}
